feat(#690): move trigram_fast.py from core/ to lib/ (Rust wrapper utility)
trigram_fast.py is a thin Python wrapper around nexus_fast Rust extension,
same pattern as hash_fast/glob_fast/grep_fast/lock_fast already in lib/.
Fix 6x type: ignore[no-redef] with _imported_* intermediate pattern.
