+++ src/nexus/search/manifest.py
@@ -67,7 +67,7 @@ def verify_imports() -> dict[str, bool]:
         "nexus.search.bm25s_search",
         "nexus.search.zoekt_client",
         "nexus.search.graph_store",
-        "nexus.core.trigram_fast",
+        "nexus.lib.trigram_fast",
     ]:
         try:
             importlib.import_module(mod)
