+++ tests/e2e/self_contained/test_trigram_search_integration.py
@@ -11,7 +11,7 @@ from unittest.mock import patch
 
 import pytest
 
-from nexus.core import trigram_fast
+from nexus.lib import trigram_fast
 from nexus.search.strategies import (
     GREP_TRIGRAM_THRESHOLD,
     SearchStrategy,
